requirements.txt
pyautogui
pynput
pystray
pillow 